project_locus/src/main/java/com/projectlocus/webservice/resources/PersonResource.java
package com.projectlocus.webservice.resources;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.projectlocus.webservice.entities.Person;
import com.projectlocus.webservice.services.PersonService;

@RestController
@RequestMapping(value = "/persons")
public class PersonResource {
	
	@Autowired
	private PersonService service;
	
	@GetMapping
	public ResponseEntity<List<Person>> findAll(){
		List<Person> p = service.findAll();
		return ResponseEntity.ok().body(p);
	}

}
